+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # Change Log
 This project adheres to Semantic Versioning
 
+## [0.1.2] - 2017-06-12
+### Added
+- Add --print-languages and --print-licenses which print a list of useful items
+  that can be used.
+
+### Changed
+- Change filetype regexes and comment types to use one internal table based on
+  language name.
+
+### Fixed
+- Fix some errors in documentation.
+
+### Added
+- Add new search-replace feature for the license. This replaces the options
+  --license-search and --license-replace from before. Now it can do multiple
+  searches.
+
 ## [0.1.1] - 2017-06-09
 ### Fixed
 - Fix numerous errors in README, including incorrect executable name and